Move SignerIdent, new package to hold account impl separate from web code
