+++ __mocks__/@thecointech/accounts/index.ts
@@ -3,7 +3,7 @@ import testWallet from './testAccount1.json';
 import Thisismy from './Thisismy.wallet.json';
 import { Wallet } from 'ethers';
 import { AccountName, getSigner } from '@thecointech/signers';
-import { TheSigner } from '@thecointech/shared/SignerIdent';
+import { TheSigner } from '@thecointech/utilities/SignerIdent';
 import { ConnectContract } from '../contract';
 import { MockIDX } from '../idx';
