+++ apps/site-app/src/containers/AddAccount/Connect/index.tsx
@@ -6,7 +6,7 @@ import { WarnIfDisabled } from './WarnIfDisabled';
 import { NameInput } from '../NewBaseClass/NameInput';
 import { ReferralInput, registerReferral } from '../NewBaseClass/ReferralInput';
 import messages from '../messages';
-import { TheSigner } from '@thecointech/shared/SignerIdent';
+import { TheSigner } from '@thecointech/utilities/SignerIdent';
 import { useHistory } from 'react-router';
 import { ButtonPrimary } from '@thecointech/site-base/components/Buttons';
 import { IAccountStoreAPI, useAccountStoreApi } from '@thecointech/shared/containers/AccountMap';
