+++ apps/site-app/src/containers/Settings/StorageOptions/download/index.tsx
@@ -3,7 +3,7 @@ import FileSaver from 'file-saver';
 import { getStoredAccountData } from '@thecointech/shared/utils/storageSync';
 import { FormattedMessage } from 'react-intl';
 import messages from './messages';
-import { isWallet } from '@thecointech/shared/SignerIdent';
+import { isWallet } from '@thecointech/utilities/SignerIdent';
 import { ButtonSecondary } from '@thecointech/site-base/components/Buttons';
