+++ apps/site-nft/src/components/AddAccount/Store/Providers/Google/StoreGoogle.tsx
@@ -4,7 +4,7 @@ import React, { useState, useEffect, useCallback } from 'react';
 import { getStoredAccountData } from '@thecointech/shared/utils/storageSync';
 import messages from './messages';
 import { FormattedMessage } from 'react-intl';
-import { isWallet } from '@thecointech/shared/SignerIdent';
+import { isWallet } from '@thecointech/utilities/SignerIdent';
 import { onInitiateLogin, setupCallback, UploadState, doSetup } from './googleUtils';
 import { useActiveAccount } from '@thecointech/shared/containers/AccountMap';
 import { Props as MessageProps } from '@thecointech/site-base/components/MaybeMessage';
