+++ libs/account/.vscode/launch.json
@@ -1,0 +1,24 @@
+{
+	// Use IntelliSense to learn about possible attributes.
+	// Hover to view descriptions of existing attributes.
+	// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+	"version": "0.2.0",
+	"configurations": [
+	{
+		"type": "pwa-node",
+		"name": "vscode-jest-tests",
+		"request": "launch",
+		"program": "${workspaceFolder}/../../node_modules/jest/bin/jest",
+		"args": [
+			"--runInBand",
+			"--config",
+			"jest.config.js"
+		],
+    "envFile": "/TheCoinEnvironments/prod.env",
+		"cwd": "${workspaceFolder}",
+		"console": "integratedTerminal",
+		"internalConsoleOptions": "neverOpen",
+    "customDescriptionGenerator": "function (def) { if (this.toString) { const _v = this.toString(); if (_v.indexOf(\"[object Object]\") < 0) return _v; } return def; }"
+	}
+	]
+}
