+++ libs/account/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@thecointech/account",
+  "version": "0.2.7",
+  "description": "TheCoin: Account definition & storage",
+  "keywords": [
+    "TheCoin",
+    "S&P500"
+  ],
+  "author": "Stephen Taylor <stephen@thecoin.io>",
+  "homepage": "http://thecoin.io",
+  "license": "GPL-3.0-or-later",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/thecointech/thecoin.git",
+    "directory": "libs/account"
+  },
+  "publishConfig": {
+    "registry": "https://npm.pkg.github.com/"
+  },
+  "scripts": {
+    "prepublishOnly": "rimraf ./build/.buildinfo",
+    "build": "yarn tsc --project tsconfig.build.json",
+    "clean": "rimraf ./build && rimraf ./coverage",
+    "test": "jest --coverage",
+    "test:summary": "jest --coverage > ./coverage/summary.txt"
+  },
+  "files": [
+    "build/**/*"
+  ],
+  "exports": "./build/index.js",
+  "types": "./build/index.d.ts",
+  "dependencies": {
+    "@thecointech/logging": "*",
+    "@thecointech/tx-blockchain": "*",
+    "@thecointech/utilities": "*"
+  },
+  "devDependencies": {
+    "jest": "*"
+  }
+}
