+++ libs/account/src/details.ts
