+++ libs/account/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../tsconfig.libs.json",
+  "compilerOptions": {
+    "outDir": "./build/",
+    "tsBuildInfoFile": "./build/.buildinfo",
+    "rootDir": "./src/",
+    "baseUrl": "./src/"
+  },
+  "include": ["src/**/*"],
+}
