+++ libs/shared/src/containers/Account/reducer.ts
@@ -4,14 +4,15 @@ import { call, StrictEffect } from 'redux-saga/effects';
 import { useInjectReducer, useInjectSaga } from "redux-injectors";
 import { IsValidAddress, NormalizeAddress } from '@thecointech/utilities';
 import { useDispatch } from 'react-redux';
-import { AccountState, DecryptCallback, DefaultAccountValues, IActions } from './types';
+import { DecryptCallback, IActions } from './types';
 import { buildSagas } from './actions';
 import { actions as FxActions } from '../../containers/FxRate/reducer';
 import { TheCoinReducer } from '../../store/immerReducer';
-import { isSigner, TheSigner } from '../../SignerIdent';
+import { isSigner, TheSigner } from '@thecointech/utilities/SignerIdent';
 import { loadAndMergeHistory, calculateTxBalances, Transaction } from '@thecointech/tx-blockchain';
 import { connectIDX } from '@thecointech/idx';
-import { AccountDetails, loadDetails, setDetails } from '../AccountDetails';
+import { AccountDetails, AccountState, DefaultAccountValues } from '@thecointech/account';
+import { loadDetails, setDetails } from '../AccountDetails';
 import { DateTime } from 'luxon';
 import { log } from '@thecointech/logging';
 import { createActionCreators, createReducerFunction } from 'immer-reducer';
