+++ libs/shared/src/containers/AccountDetails/index.ts
@@ -1,2 +1,1 @@
-export * from './types';
 export * from './persist';
