+++ libs/shared/src/containers/AccountMap/index.tsx
@@ -1,4 +1,3 @@
 export * from "./reducer";
 export * from "./types";
 export * from "./selectors";
-export type { AccountState } from '../Account/types';
