+++ libs/tx-blockchain/src/types.ts
@@ -1,7 +1,6 @@
 //
 // All public-facing types are defined here.
 //
-// TODO: dedup with shared/account/types
 // TODO: convert to DateTime / Decimal for appropriate entries
 
 import { DateTime } from "luxon";
